feat: creates import module to get import_path
